refactor: :recycle: desktop | tablet 의 공통된 부분을 common폴더 내의 aboveTablet으로 빼놓음
