+++ app/src/components/elements/NavBar/Tablet/index.tsx
@@ -1,23 +1,13 @@
-import { AppLogoTitleButton } from '@/components/elements/AppLogoTitleButton';
-import { DesktopNavMenu } from '../Desktop/DesktopNavMenu';
 import styled from '@emotion/styled';
-import { MouseEvent, useEffect } from 'react';
-import { VStack, Overlay, Button } from '@/components/common';
-import { NavProfile } from '@/components/elements/NavProfile';
+import { useEffect } from 'react';
+import { Overlay } from '@/components/common';
 import { isNavBarOpenAtom } from '@/utils/atoms/isNavBarOpenAtom';
-import { useAtom, useAtomValue } from 'jotai';
-import { userAtom } from '@/utils/atoms/userAtom';
-import { useNavigate } from 'react-router-dom';
+import { useAtom } from 'jotai';
+import { DesktopNavBarLayout } from '../Desktop';
+import { AboveTabletNavBar } from '../common/AboveTabletNavBar';
 
 export const TabletNavBar = () => {
   const [isNavBarOpen, setIsNavBarOpen] = useAtom(isNavBarOpenAtom);
-  const user = useAtomValue(userAtom);
-  const navigate = useNavigate();
-
-  const handleClick = (e: MouseEvent<HTMLElement>) => {
-    e.preventDefault();
-    navigate(`/profile/${user.login}`);
-  };
 
   useEffect(() => {
     setIsNavBarOpen(false);
@@ -30,20 +20,7 @@ export const TabletNavBar = () => {
         <Overlay onClick={() => setIsNavBarOpen((cur) => !cur)} />
       )}
       <TabletNavBarLayout isOpen={isNavBarOpen}>
-        <VStack h="100%" spacing="6rem">
-          <AppLogoTitleButton />
-          <Button
-            onClick={handleClick}
-            element={
-              <NavProfile
-                imageUrl={user.imageUrl}
-                name={user.name}
-                login={user.login}
-              />
-            }
-          />
-          <DesktopNavMenu />
-        </VStack>
+        <AboveTabletNavBar />
       </TabletNavBarLayout>
     </>
   );
@@ -53,15 +30,7 @@ type TabletNavBarLayoutProps = {
   isOpen: boolean;
 };
 
-const TabletNavBarLayout = styled.nav<TabletNavBarLayoutProps>`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 24rem;
-  height: 100%;
-  padding: 3rem;
-  background-color: ${({ theme }) => theme.colors.mono.white};
-  border-right: 0.1rem solid ${({ theme }) => theme.colors.mono.gray.light};
+const TabletNavBarLayout = styled(DesktopNavBarLayout)<TabletNavBarLayoutProps>`
   transform: ${({ isOpen }) => !isOpen && 'translateX(-100%)'};
   transition: all 0.3s;
   z-index: 3;
