+++ app/src/components/elements/NavBar/common/NavMenuIcon.tsx
