+++ app/src/components/elements/NavBar/common/NavProfile.tsx
